feat(editor): add inline camera capture and video playback for journal entries
This commit introduces a complete camera and video capture system to the entry
editor, enabling users to capture photos and videos directly within journal
entries. The implementation follows a clean architecture approach with proper
separation between the capture infrastructure and UI components.

The camera capture system is built around a CameraCaptureManager interface that
abstracts platform-specific camera operations. On Android, this is implemented
using CameraX with support for both ImageCapture and VideoCapture use cases,
complete with proper lifecycle management and MediaStore integration for
persisting captured media. The iOS and Desktop implementations are provided as
stubs that can be filled in as those platforms mature.

A new CameraViewModel manages the capture state, handling transitions between
photo and video modes, tracking recording duration, and coordinating with the
capture manager. The UI state is modeled through CameraUiState which includes
formatted duration display for video recordings.

The editor UI has been updated to support camera blocks through CameraBlockEditor,
which displays an inline camera preview that expands to fullscreen when tapped.
Users can switch between front and back cameras, toggle between photo and video
modes, and capture media with a simple shutter button. Captured content displays
with a thumbnail and optional caption field.

Video playback on Android uses Media3/ExoPlayer for production-quality playback.
The VideoPlayerContent component shows a thumbnail with play button overlay when
paused, transitioning to full video playback on tap. Proper lifecycle management
ensures playback pauses when the app backgrounds.

The EmptyEditorStateContent has been reorganized into a 2x2 grid layout that
prominently features camera capture alongside text, audio, and gallery options.
This makes the camera capture feature immediately discoverable to users.

Several bug fixes are included: LiveCameraPreview now properly binds to the
lifecycle (previously the camera was never actually started), and the
DesktopAudioPlaybackManager has a corrected type conversion for frame length.

The entry persistence layer has been updated to handle CameraBlockUiState,
converting captured media to either Image or Video JournalNote types based on
the mediaType field. This ensures captured content is properly saved and can
be restored when editing existing entries.

Test coverage includes unit tests for CameraUiState, CameraBlockUiState, and
VideoBlockUiState, verifying state initialization, duration formatting, and
data class behavior. A FakeCameraCaptureManager is provided for testing
components that depend on camera functionality.

+++ client/feature/editor/src/desktopMain/kotlin/app/logdate/feature/editor/ui/audio/DesktopAudioPlaybackManager.kt
@@ -75,7 +75,7 @@ class DesktopAudioPlaybackManager : AudioPlaybackManager {
             clip = newClip
             
             // Get total frames for progress calculation
-            totalFrames = newClip.frameLength
+            totalFrames = newClip.frameLength.toLong()
             
             // Add listener for playback completion
             newClip.addLineListener { event ->
